Issue #6728 - QUIC and HTTP/3
- QPACK exceptions now use long instead of int, to be consistent with other error codes.
- Fixed ManagedSelector to count down the stop latches in finally blocks, so that they are always counted down even in case of exceptions.
- Improved exception handling in case of closes.

+++ jetty-http3/http3-common/src/main/java/org/eclipse/jetty/http3/internal/parser/BodyParser.java
@@ -86,7 +86,7 @@ public abstract class BodyParser
         }
     }
 
-    protected void notifyStreamFailure(long streamId, int error, Throwable failure)
+    protected void notifyStreamFailure(long streamId, long error, Throwable failure)
     {
         try
         {
